semantify/test_semantify.py
import argparse
import os
import unittest
from unittest.mock import patch, MagicMock
from semantify.semantify import Semantify, main
from semantify.interfaces import IMetadataManager, IRecommender
from semantify.utils.frontmatter_utils import FrontmatterUtils
from semantify.recommender.blog_recommender import BlogRecommender
from semantify.summarizer.qa_generator import QAGenerator
from semantify.services.sqlite import SQLiteDBService
from semantify.services.openai import OpenAIService


class MockMetadataManager(IMetadataManager):
    def read_metadata(self, mdx_path):
        # Mock implementation
        pass

    def update_metadata(self, mdx_path, metadata, content):
        # Mock implementation
        pass

    def clear_sections(self, blog_directory, sections):
        # Mock implementation
        pass


class MockRecommender(IRecommender):
    def generate_recommendations(self, slug):
        # Mock implementation
        return ['post1', 'post2', 'post3']

    def clear_embeddings(self):
        # Mock implementation
        pass


class TestSemantify(unittest.TestCase):
    def setUp(self):
        self.mock_metadata_manager = MagicMock(spec=MockMetadataManager)
        self.mock_recommender = MagicMock(spec=MockRecommender)
        self.mock_qa_generator = MagicMock(spec=QAGenerator)
        self.blog_directory = "/path/to/blog/directory"
        self.semantify_instance = Semantify(
            metadata_manager=self.mock_metadata_manager,
            blog_directory=self.blog_directory,
            recommender=self.mock_recommender,
            qa_generator=self.mock_qa_generator
        )

    @patch('os.path.exists', return_value=True)
    def test_process_blog_post(self, mock_exists):
        mdx_path = "path/to/blog_post.mdx"
        self.mock_metadata_manager.read_metadata.return_value = (
            {'slug': 'test_slug'}, "content", None)
        self.mock_qa_generator.generate_qa_summary.return_value = "QA Summary"

        self.semantify_instance.process_blog_post(mdx_path)

        self.mock_metadata_manager.read_metadata.assert_called_once_with(
            mdx_path)
        self.mock_qa_generator.generate_qa_summary.assert_called_once()
        self.mock_recommender.generate_recommendations.assert_called_once()
        self.mock_metadata_manager.update_metadata.assert_called()

    @patch('argparse.ArgumentParser.parse_args')
    def test_start(self, mock_parse_args):
        mock_parse_args.return_value = argparse.Namespace(
            blog_directory='/path/to/blog/directory',
            openai_api_key='your-api-key',
            replace_reading_time=False,
            replace_recommendations=False,
            replace_qa=False
        )

        with patch('semantify.semantify.os.walk') as mock_walk:
            mock_walk.return_value = [
                ('/path/to/blog/directory', [], ['test_post1.mdx', 'test_post2.mdx'])]
            with patch.object(self.semantify_instance, 'process_blog_post') as mock_process_blog_post:
                self.semantify_instance.start(mock_parse_args.return_value)
                mock_process_blog_post.assert_called()


if __name__ == "__main__":
    unittest.main()
